+++ .github/CHANGELOG.md
@@ -4,7 +4,9 @@ This file contains update logs for this project. The top may contain a `Unreleas
 ---
 ## Unreleased
 ### Changed
+* Added config options `commands-on-maintenance-enable` and `commands-on-maintenance-disable` to define commands to be executed after maintenance has been toggled
 * Added config option `enable-pingmessages` to can change whether a custom maintenance motd should be displayed during maintenance
+* Added language fields `scheduletimerBroadcast` and `singleScheduletimerBroadcast` with the variables `%SERVER%`, `%TIME%`, and `%DURATION%` instead of using the normal starttimer broadcasts
 ---
 ## 3.0.6 - February 17th 2021
 ### Changed